requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
alembic==1.13.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
